+++ packages/terra-clinical-site/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+Lock wepback-dev-server at last version supporting IE10 (1.7.1)
 
 1.3.0 - (August 16, 2017)
 -----------------