+++ packages/terra-clinical-site/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Lock webpack-dev-server at last version supporting IE10 (1.7.1)
+* Uplift site header to use collapsible menu view
 * Updated webpack config to align with webpack 3 documentation
 
 1.3.0 - (August 16, 2017)